parsing shoulder from identifiers
I went with the approach of fixing this in DCM because two datasets can have the same doi after the shoulder but different shoulders. Even if practially that won't happen at this point we should probably update DCM to do a check on the full doi.

This fix is still unclean with comment code